Adds missing Identity Window API endpoints; Makes Identity configurable
